requirements/prd.txt
mysql-connector-python==8.0.29
pandas==1.4.3
python-dotenv==0.20.0